requirements.txt
numpy>=1.26
pandas>=2.1
requests>=2.31
python-dateutil>=2.8
python-dotenv>=1.0
pytz>=2023.3
numba>=0.58
xgboost>=1.7
scikit-learn==1.7.2
